requirements.txt
fastapi==0.79.0
numpy==1.21.5
opencv_python==4.6.0.66
Pillow==9.2.0
pydantic==1.9.1
pytesseract==0.3.9
python-multipart==0.0.5
uvicorn==0.18.2
